refactor(frontend): 미디어 타입을 entities 로 통합
