+++ frontend/src/entities/media/index.ts
@@ -1,4 +1,4 @@
 export { getPhotos } from "./api/getPhotos";
 export { usePhotosQuery } from "./model/usePhotosQuery";
-export type { MediaItem, MediaList, PhotoFilter } from "./model/types";
+export type { Media, MediaItem, MediaList, MediaStatus, PhotoFilter } from "./model/types";
 export { MediaCard } from "./ui/MediaCard";
